backend/requirements.txt
asgiref==3.7.2
asn1crypto==1.5.1
base58==2.1.1
bit==0.8.0
bitcoin==1.1.39
bitcoinlib==0.6.11
blockcypher==1.0.93
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
coincurve==18.0.0
cryptography==41.0.3
Django==4.2.4
django-cors-headers==4.2.0
django-extensions==3.2.3
djangorestframework==3.14.0
dulwich==0.21.5
ecdsa==0.18.0
greenlet==2.0.2
idna==3.4
MarkupSafe==2.1.3
numpy==1.25.2
pycoin==0.92.20230326
pycparser==2.21
pycryptodome==3.18.0
pyOpenSSL==23.2.0
python-dateutil==2.8.2
python-decouple==3.8
python-git-secrets==0.1.3
pytz==2023.3
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.20
sqlparse==0.4.4
typing-extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
werkzeug==2.3.7
